+++ components.d.ts
@@ -4,6 +4,7 @@
 
 declare module 'vue' {
   export interface GlobalComponents {
+    Color: typeof import('./src/components/Color.vue')['default']
     Command: typeof import('./src/components/Command.vue')['default']
     ElAutocomplete: typeof import('element-plus/es')['ElAutocomplete']
     ElButton: typeof import('element-plus/es')['ElButton']