+++ src/commands/index.ts
@@ -6,3 +6,4 @@ export * as invoke from "./invoke";
 export * as lorem from "./lorem";
 export * as sort from "./sort";
 export * as emoji from "./emoji";
+export * as bg from "./bg";